requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.4
dj-database-url==3.0.1
Django==4.2.27
packaging==26.0
pdf2image==1.17.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyMySQL==1.1.2
pytesseract==0.3.13
sqlparse==0.5.5
typing_extensions==4.15.0
